Reimplemented logic back into fragment.  Beginning testing
